indexer-hukuk/requirements.txt
pdfplumber~=0.11.0
openai~=1.43.1
langchain~=0.1.14
python-dotenv~=1.0.0
uvicorn~=0.28.0
fastapi~=0.110.0
pydantic~=2.6.3
azure-storage-blob~=12.18.0
azure-search-documents~=11.4.0
